+++ rampt/steps/analysis/summary_pipe.py
@@ -504,7 +504,7 @@ class Summary_Runner(Pipe_Step):
         )
 
         # Search for relevant files
-        matched_in_paths = in_paths.copy()
+        matched_in_paths = {}
         for file_type, path in in_paths.items():
             path = to_list(path)[0]
             if file_type in self.data_ids["in_paths"]: