+++ rampt/steps/annotation/sirius_pipe.py
@@ -274,7 +274,7 @@ class Sirius_Runner(Pipe_Step):
                     config = join(in_paths, entry)
 
         # Search for relevant files
-        matched_in_paths = in_paths.copy()
+        matched_in_paths = {}
         for file_type, path in in_paths.items():
             path = to_list(path)[0]
             if file_type in self.data_ids["in_paths"]: